Fix tests and have better grouping
